Move MockFileSystem to test/helpers
