+++ test/lsp/ProtocolTest.cc
@@ -1,53 +1,10 @@
 #include "test/lsp/ProtocolTest.h"
+#include "test/helpers/MockFileSystem.h"
 #include "test/helpers/lsp.h"
 #include "test/helpers/position_assertions.h"
 
 namespace sorbet::test::lsp {
 using namespace std;
-
-MockFileSystem::MockFileSystem(std::string_view rootPath) : rootPath(string(rootPath)) {}
-
-string makeAbsolute(string_view rootPath, string_view path) {
-    if (path[0] == '/') {
-        return string(path);
-    } else {
-        return fmt::format("{}/{}", rootPath, path);
-    }
-}
-
-void MockFileSystem::writeFiles(const vector<pair<string, string>> &initialFiles) {
-    for (auto &pair : initialFiles) {
-        writeFile(pair.first, pair.second);
-    }
-}
-
-string MockFileSystem::readFile(string_view path) const {
-    auto file = contents.find(makeAbsolute(rootPath, path));
-    if (file == contents.end()) {
-        throw FileNotFoundException();
-    } else {
-        return file->second;
-    }
-}
-
-void MockFileSystem::writeFile(string_view filename, string_view text) {
-    contents[makeAbsolute(rootPath, filename)] = text;
-}
-
-void MockFileSystem::deleteFile(string_view filename) {
-    auto file = contents.find(makeAbsolute(rootPath, filename));
-    if (file == contents.end()) {
-        throw FileNotFoundException();
-    } else {
-        contents.erase(file);
-    }
-}
-
-vector<string> MockFileSystem::listFilesInDir(string_view path, const UnorderedSet<string> &extensions, bool recursive,
-                                              const std::vector<std::string> &absoluteIgnorePatterns,
-                                              const std::vector<std::string> &relativeIgnorePatterns) const {
-    Exception::raise("Not implemented.");
-}
 
 void ProtocolTest::SetUp() {
     rootPath = "/Users/jvilk/stripe/pay-server";
