+++ test/lsp/ProtocolTest.h
@@ -3,8 +3,8 @@
 
 #include "gtest/gtest.h"
 // ^ Violates linting rules, so include first.
-#include "common/FileSystem.h"
 #include "main/lsp/wrapper.h"
+#include "test/helpers/MockFileSystem.h"
 
 namespace sorbet::test::lsp {
 using namespace sorbet::realmain::lsp;
@@ -13,22 +13,6 @@ struct ExpectedDiagnostic {
     std::string path;
     int line;
     std::string message;
-};
-
-class MockFileSystem final : public FileSystem {
-private:
-    UnorderedMap<std::string, std::string> contents;
-    std::string rootPath;
-
-public:
-    MockFileSystem(std::string_view rootPath);
-    void writeFiles(const std::vector<std::pair<std::string, std::string>> &files);
-    std::string readFile(std::string_view path) const override;
-    void writeFile(std::string_view filename, std::string_view text) override;
-    void deleteFile(std::string_view filename);
-    std::vector<std::string> listFilesInDir(std::string_view path, const UnorderedSet<std::string> &extensions,
-                                            bool recursive, const std::vector<std::string> &absoluteIgnorePatterns,
-                                            const std::vector<std::string> &relativeIgnorePatterns) const override;
 };
 
 class ProtocolTest : public testing::Test {
@@ -44,7 +28,7 @@ protected:
     // Currently active diagnostics, specifically using map to enforce sort order on filename.
     std::map<std::string, std::vector<std::unique_ptr<Diagnostic>>> diagnostics;
     // Emulated file system.
-    std::shared_ptr<MockFileSystem> fs;
+    std::shared_ptr<sorbet::test::MockFileSystem> fs;
 
     /** The next ID to use when sending an LSP message. */
     int nextId = 0;
